+++ src/compiler/MethodGenerationContext.h
@@ -66,7 +66,7 @@ public:
     void SetPrimitive(bool prim = true);
 
     uint8_t AddLiteral(vm_oop_t lit);
-    int8_t AddLiteralIfAbsent(vm_oop_t lit);
+    uint8_t AddLiteralIfAbsent(vm_oop_t lit);
     void UpdateLiteral(vm_oop_t oldValue, uint8_t index, vm_oop_t newValue);
 
     void MarkFinished();
@@ -93,7 +93,7 @@ public:
     void RemoveLastPopForBlockLocalReturn();
 
     uint8_t GetNumberOfArguments();
-    void AddBytecode(uint8_t bc, size_t stackEffect);
+    void AddBytecode(uint8_t bc, int64_t stackEffect);
     void AddBytecodeArgument(uint8_t bc);
     size_t AddBytecodeArgumentAndGetIndex(uint8_t bc);
