+++ src/primitives/System.cpp
@@ -88,7 +88,7 @@ static vm_oop_t sysLoad_(vm_oop_t /*unused*/, vm_oop_t rightObj) {
 
 static vm_oop_t sysExit_(vm_oop_t /*unused*/, vm_oop_t err) {
     int64_t const err_no = INT_VAL(err);
-    Quit(err_no);
+    Quit((int32_t)err_no);
 }
 
 static vm_oop_t sysPrintString_(vm_oop_t leftObj, vm_oop_t rightObj) {
